refactor(remote)!: move remote access primitives into core and ui
RemoteAccess and BrowserToken move to Lattice\Core\Remote, the abstract
RemoteComponent base moves to Lattice\Ui\Components. RemoteComponent now
resolves the new core contract ResolvesRemoteSourceEndpoints instead of the
framework's RemoteSourceRegistry, so remote-capable components only need
core and ui at the class level. The registry, source definitions, and the
schema resolver stack stay in the framework, which binds the contract.

Core's hand-written RemoteAccess/BrowserToken TS types in api.ts are now
re-exports of the generated wire types.

BREAKING CHANGE: Lattice\Remote\RemoteAccess and Lattice\Remote\BrowserToken
are now Lattice\Core\Remote\RemoteAccess and Lattice\Core\Remote\BrowserToken;
Lattice\Remote\Components\RemoteComponent is now
Lattice\Ui\Components\RemoteComponent. RemoteAccess and BrowserToken moved
from @lattice-php/lattice/types/generated to @lattice-php/core.
