+++ packages/core/src/Contracts/ResolvesRemoteSourceEndpoints.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace Lattice\Core\Contracts;
+
+interface ResolvesRemoteSourceEndpoints
+{
+    public function endpointFor(string $source): string;
+}
